+ fix (errors occured when trying to get non-existing favourite_product)

+++ routes/products.py
@@ -32,7 +32,7 @@ def get_favourite_products(current_user: models.User = Depends(get_current_user)
 
     return db_items.all()
 
-@products_router.get("/favourites/{product_id}", response_model=schemas.FavouriteProduct)
+@products_router.get("/favourites/{product_id}", response_model=schemas.FavouriteProduct|None)
 def get_favourite_product(product_id: int, current_user: models.User = Depends(get_current_user), db: Session = Depends(get_db)):
     db_item = db.query(models.FavouriteProduct).filter(models.FavouriteProduct.product_id == product_id).first()
 
@@ -45,7 +45,7 @@ def add_favourite_products(product_id: int = Form(...), current_user: models.Use
         product_id = product_id,
         user_id = current_user.id
     )
-
+    
     db.add(db_item)
     db.commit()
     db.refresh(db_item)
@@ -54,12 +54,12 @@ def add_favourite_products(product_id: int = Form(...), current_user: models.Use
 
 @products_router.delete("/favourites/{product_id}", response_model=bool)
 def delete_favourite_product(product_id: int, current_user: models.User = Depends(get_current_user), db: Session = Depends(get_db)):
-    db_item = db.query(models.FavouriteProduct).filter(models.FavouriteProduct.id != product_id).first()
-
-    db.delete(db_item)
-    db.commit()
-
-    return True
+    db_item = db.query(models.FavouriteProduct).filter(models.FavouriteProduct.product_id == product_id).first()
+    if(db_item != None):
+        db.delete(db_item)
+        db.commit()
+        return True
+    return False
 
 @products_router.get("/recent", response_model=List[schemas.ProductHistory])
 def get_recently_used_products(current_user: models.User = Depends(get_current_user), db: Session = Depends(get_db)):
